CUDA RDC: Improve handling of object library and CUDA_RUNTIME_LIBRARY (#1696)
* Increment CudaRdcUtils version to 5

* Explicitly add/pass along only RDC libraries

* Export CUDA_RUNTIME_LIBRARY for RDC libraries

* Improve CUDA_RUNTIME_LIBRARY determination (for export libs)

In case a library does not have a CUDA_RUNTIME_LIBRARY
(i.e. case of exported library with CudaRdcUtils v4 or older)
we now 'guess' the type by checking whether the middle and
object library are the same, if they are this is a static library
and it "probably" require the static CUDA runtime.

* Improve CUDA_RUNTIME_LIBRARY determination (for mixed builds)

In case of mixed builds (local static libraries trying to use
installed shared library), the local library need to use the
CUDA runtime used by the dependent/installed library (since
we can't rebuild them to have any other one).

However if we are trying to link against 2 libraries with
different CUDA runtime, we need to fail.

* Increment CudaRdcUtils version to 6